chore: move std_malloc_resource under fb2
Also, remove the deprecated fiber2.h file.
